@@ -1,12 +1,10 @@
 package edu.cascadia.mobile.apps.passwordcheck;
 
-import android.arch.lifecycle.ViewModel;
-import android.arch.lifecycle.ViewModelProvider;
-import android.databinding.DataBindingUtil;
-import android.support.v7.app.AppCompatActivity;
+import androidx.lifecycle.ViewModel;
+import androidx.lifecycle.ViewModelProvider;
+import androidx.databinding.DataBindingUtil;
+import androidx.appcompat.app.AppCompatActivity;
 import android.os.Bundle;
-
-import edu.cascadia.mobile.apps.passwordcheck.databinding.ActivityMainBinding;
 
 public class MainActivity extends AppCompatActivity {
 
